[100th Commit] Updates on several files.
Made seperate string for players_txt to be used to select a player map. character selction is not done tho. Made it so that we can customize what cards a player could have. Made a thing that makes route map contains specific enemies to be played with. each enemy will have specific cards to be customized with. So mannnyy stuuufffs iddkkkkk anymmoooree